\docType{package}
\name{cloudRmpiJars}
\alias{cloudRmpiJars}
\alias{cloudRmpiJars-package}
\title{Third-party jars for cloudRmpi}
\description{
  Cloud-based MPI Parallel Processing for R (cloudRmpi)
}
\details{
  \tabular{ll}{ Package: \tab cloudRmpiJars\cr Type: \tab
  Package\cr Version: \tab 1.1--\cr Date: \tab
  2012-04-18\cr License: \tab GPL (>= 3)\cr }

  This package contains third-party Java class jars used by
  cloudRmpi. cloudRmpi is means for doing parallel
  processing in R, using MPI on a cloud-based network.
}
\author{
  Barnet Wagman \email{bw@norbl.com}
}
\keyword{package}

